Move test cases to new component
