+++ src/app/components/players-hand/players-hand.component.spec.ts
@@ -2,9 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatDialogModule } from '@angular/material/dialog';
 import { PlayersHandComponent } from './players-hand.component';
 import { PlayerDisplayNamePipe } from '../../pipes/player-display-name/player-display-name.pipe';
-
-const USERNAME_ONE = 'Custom name 1';
-const USERNAME_TWO = 'Custom name 2';
+import { CommunicationPositionPipe } from 'src/app/pipes/communication-position/communication-position.pipe';
 
 describe('PlayersHandComponent', () => {
   let component: PlayersHandComponent;
@@ -12,7 +10,7 @@ describe('PlayersHandComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [PlayersHandComponent],
+      declarations: [PlayersHandComponent, CommunicationPositionPipe],
       imports: [MatDialogModule],
       providers: [PlayerDisplayNamePipe],
     }).compileComponents();
@@ -26,58 +24,5 @@ describe('PlayersHandComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
-  });
-
-  it('should prevent dragging to players hand if they have communicated', () => {
-    const communicationCard = {
-      suit: 'pink',
-      value: 2,
-      playerPosition: 1,
-      username: USERNAME_TWO,
-    };
-    component.communicationCard = [communicationCard];
-    component.revealedCommunications = [
-      {
-        type: 'highest',
-        card: communicationCard,
-      },
-      {
-        type: 'lowest',
-        card: {
-          suit: 'green',
-          value: 3,
-          playerPosition: 2,
-          username: USERNAME_TWO,
-        },
-      },
-    ];
-
-    const expected = 'playing-mat';
-    const actual = component.communicationDragTo();
-    expect(actual).toBe(expected);
-  });
-
-  it('should allow dragging to players hand if they have not communicated', () => {
-    const communicationCard = {
-      suit: 'pink',
-      value: 2,
-      playerPosition: 1,
-      username: USERNAME_ONE,
-    };
-    component.communicationCard = [communicationCard];
-    component.revealedCommunications = [
-      {
-        type: 'lowest',
-        card: {
-          suit: 'green',
-          value: 3,
-          playerPosition: 2,
-          username: USERNAME_ONE,
-        },
-      },
-    ];
-    const expected = ['playing-mat', 'players-hand'];
-    const actual = component.communicationDragTo();
-    expect(actual).toEqual(expected);
   });
 });
